requirements.txt
ruamel.yaml~=0.17
pyoxidizer~=0.22